NEWTAG-624 Implement WieldCategory CodeControl
Add WieldCategory so it can be a format
