+++ code/src/itest/plugin/lsttokens/editcontext/equipment/WeildIntegrationTest.java
@@ -1,16 +1,16 @@
 /*
  * Copyright (c) 2007 Tom Parker <thpr@users.sourceforge.net>
- *
+ * 
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
  * Software Foundation; either version 2.1 of the License, or (at your option)
  * any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
- *
+ * 
  * You should have received a copy of the GNU Lesser General Public License
  * along with this library; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
@@ -35,7 +35,7 @@ import plugin.pretokens.writer.PreVariableWriter;
 import org.junit.jupiter.api.Test;
 import util.GameModeSupport;
 
-public class WieldIntegrationTest extends
+public class WeildIntegrationTest extends
 		AbstractIntegrationTestCase<Equipment>
 {
 	private static final CDOMPrimaryToken<Equipment> token = new WieldToken();
